Move package of ShardingSphereServerException
