+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-spring/shardingsphere-sharding-spring-namespace/src/main/java/org/apache/shardingsphere/sharding/spring/namespace/exception/UnsupportedShardingAlgorithmTypeServerException.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.sharding.spring.namespace.exception;
 
-import org.apache.shardingsphere.infra.util.exception.external.ShardingSphereServerException;
+import org.apache.shardingsphere.infra.util.exception.external.server.ShardingSphereServerException;
 
 /**
  * Unsupported sharding algorithm type exception.
