+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-spring-infra/shardingsphere-jdbc-spring-boot-starter-infra/src/main/java/org/apache/shardingsphere/spring/boot/exception/UnsupportedDataSourceTypeServerException.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.spring.boot.exception;
 
-import org.apache.shardingsphere.infra.util.exception.external.ShardingSphereServerException;
+import org.apache.shardingsphere.infra.util.exception.external.server.ShardingSphereServerException;
 
 /**
  * Unsupported data source type exception.
