+++ shardingsphere-proxy/shardingsphere-proxy-bootstrap/src/main/java/org/apache/shardingsphere/proxy/exception/DatabaseServerLoadingServerException.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.proxy.exception;
 
-import org.apache.shardingsphere.infra.util.exception.external.ShardingSphereServerException;
+import org.apache.shardingsphere.infra.util.exception.external.server.ShardingSphereServerException;
 
 import java.sql.SQLException;
